move c source to src folder
This fixes a confusing bug where importing hla in the pyhla-evolved folder
would result in it appearing that the _omt and _rti modules were empty as
they were in fact being imported as namespace packages

This now means that hla cannot be imported in the top level folder unless
python setup.py develop is run first. This is a feature, not a bug.